pyproject.toml
[tool.poetry]
name = "calendar-todoist-sync"
version = "0.1.0"
description = "Python project to synchorinize tasks between Todoist and Google Calendar."
authors = ["Eric Suarez"]
license = "GNU General Public License (GPL)"
readme = "README.md"
packages = [{include = "calendar_todoist_sync"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.111.0"
todoist-api-python = "^2.1.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
